+++ kernel/shared/src/main/scala/cats/effect/kernel/syntax/package.scala
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2020 Typelevel
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package cats.effect.kernel
+
+package object syntax {
+
+  object all extends AllSyntax
+
+  object concurrent extends ConcurrentSyntax
+  object temporal extends TemporalSyntax
+  object async extends AsyncSyntax
+  object syncEffect extends SyncEffectSyntax
+  object effect extends EffectSyntax
+}
